backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
flask-jwt-extended==4.4.4
Flask-Cors==4.0.0
python-dotenv==1.0.0
pymysql==1.1.0
werkzeug==2.2.3
SQLAlchemy==2.0.20